+++ docs/html/search/all_8.js
@@ -1,0 +1,5 @@
+var searchData=
+[
+  ['timing_20stats_30',['Timing Stats',['../md__home_brian_src_AceCommon_src_timing_stats_README.html',1,'']]],
+  ['timingstats_31',['TimingStats',['../classace__common_1_1TimingStats.html',1,'ace_common::TimingStats'],['../classace__common_1_1TimingStats.html#a8ffb99931928b622e696addddf426b96',1,'ace_common::TimingStats::TimingStats()']]]
+];