+++ docs/html/search/functions_7.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['timingstats_64',['TimingStats',['../classace__common_1_1TimingStats.html#a8ffb99931928b622e696addddf426b96',1,'ace_common::TimingStats']]]
+];