+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 * Unreleased
+* 1.1 (2020-10-29)
+    * Split `src/ace_common` to `src/arithmetic` and `src/pstrings`.
+    * Move `src/print_str` from AceUtils to AceCommon library.
+    * Move `src/print_utrils` from AceUtils to AceCommon library.
+    * Move `src/timing_stats` from AceUtils to AceCommon library.
+    * Move `src/url_encoding` from AceUtils to AceCommon library.
 * 1.0 (2020-10-20)
     * Add `src/ace_common/arithmetic.h` and unit tests
     * Add `src/ace_common/pstrings.h` and unit tests